Commit offsets according to a timer, rather than end of batch
endOfBatch as the trigger for committing offsets didn't work because
there was a race between the receipt of the AMQP disposition and the
arrival of the next batch. If the next batch arrived before the disposition
we wouldn't commit.

Committing on a periodic timer means commit will always happen,
irrespective of the timing of catches and dispositions. It also guards against
https://issues.apache.org/jira/browse/KAFKA-3806